RavenDB-13111 unify cluster tx tests
